Update version.txt
Make the version check text look for 1.0.6.0. Sorry PPA users and other not-fully-updated community packages. We need to ensure that Dalamud works too.

Ideally, this should be something cleaner, and maybe more informative beyond a string compare. I nominate Blooym to make an API we can host on Kamori for it.